Revert "more validation"
