+++ client/src/assets/main.css
@@ -19,6 +19,7 @@ body{
 
 h1 {
     color:#1D4367;
+    padding: 30px;
 }
 
 .containter {
