+++ client/src/pages/login/student.js
@@ -31,13 +31,14 @@ class StudentLogin extends Component {
             password
         })
         .then(res => {
-          const student = res.data;
-          if (student) {
-            API.getByUserID(type, student._id)
-            .then(resp => {
-              this.props.history.push(`/student/${student._id}`)
+          console.log(res);
+          if (res.data) {
+            API.getByUsername(type + "s", username)
+            .then(res => {
+              this.props.history.push(`/student/${res.data._id}`)
             });
           }
+          console.log("not found");
         })
 
     }
